Go-protein input reading routines
Implemented routines for reading in parameters for a go-protein. This
includes interaction parameters for the protein in the block '<go>' as
well as positions, velocities etc. in other blocks.

At this stage, only input reading works - the code does not yet do
anything with the go-particles.

Also, only one go-protein is allowed in the current setup: all
go-particles must be a part of a single chain.

There will be some conflicts between the go-particles and atoms. While
some things are expected to work in a simulation with both gos and
atoms, many things will not. So, it is better to run the simulations
without mixing the gos and the atoms until a better framework is
constructed.