Implemented go-force field
Implemented the force field for go-protein particles.

Also added the velocity-verlet and langevin routines for go particles,
and the continuation file writing routines. In principle, the code
should be simulation ready, but it is completely untested, so there are
surely plenty of bugs.
Especially the MPI-version has not been even test run yet.